rearrange hardware controller dir
