+++ api/src/opentrons/hardware_control/constants.py
@@ -1,0 +1,6 @@
+""" Constants for the :py:class:`.hardware_control.API` instances.
+"""
+SHAKE_OFF_TIPS_SPEED = 50  # mm/s
+SHAKE_OFF_TIPS_DROP_DISTANCE = 2.25  # mm
+SHAKE_OFF_TIPS_PICKUP_DISTANCE = 0.3  # mm
+DROP_TIP_RELEASE_DISTANCE = 20  # mm
