+++ api/src/opentrons/hardware_control/util.py
@@ -1,9 +1,10 @@
 """ Utility functions and classes for the hardware controller"""
 import functools
+import inspect
 import asyncio
 import logging
 from typing import Dict, Any, Optional, Callable
-from . import types, adapters, API, HardwareAPILike
+from . import types, adapters
 
 mod_log = logging.getLogger(__name__)
 
@@ -19,3 +20,17 @@ def use_or_initialize_loop(loop: Optional[asyncio.AbstractEventLoop]
     checked_loop = loop or asyncio.get_event_loop()
     checked_loop.set_exception_handler(_handle_loop_exception)
     return checked_loop
+
+
+def log_call(func):
+    if inspect.iscoroutinefunction(func):
+        @functools.wraps(func)
+        async def _log_call_inner(*args, **kwargs):
+            args[0]._log.debug(func.__name__)
+            return await func(*args, **kwargs)
+    else:
+        @functools.wraps(func)
+        def _log_call_inner(*args, **kwargs):
+            args[0]._log.debug(func.__name__)
+            return func(*args, **kwargs)
+    return _log_call_inner
